requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
python-dotenv
requests
PyJWT
pytest
